Eliminate polling from debugging APIs; fix full-range breakpoint overflow
Replace polling-based run_and_wait_for_stop with event-stream-based
approach: subscribe to WatchExecutionState, call run(), wait for the
running-to-stopped transition. Skip stale stopped events using
saw_running flag.

Fix run_until() to skip stale stopped events on the stream.

Fix Run RPC to enqueue "running" event with is_running=true BEFORE
calling resume(), preventing the race where a breakpoint fires before
the running event is queued.

Drain stale events from the queue when a new stream subscription opens.

Fix uint16_t overflow in BreakpointEntry.end and WatchpointEntry.end:
0x10000 truncated to 0, making full-range breakpoints [0, 0x10000)
match nothing. Changed to uint32_t.

Fix DebuggerService update methods to not truncate end_address.

Add event stream and full-range breakpoint tests:
- test_stream_delivers_breakpoint_event: verifies running+stopped events
- test_run_until_uses_stream: verifies run_until uses the stream
- test_cycle_budget_breakpoint: full-range + cycles >= target
- test_memory_predicate_breakpoint: full-range + mem[addr] == value
- C++ test for full-range breakpoint with cycle condition

Fix CoupledSystem.run_for to use server-side cycle-budget breakpoint
instead of polling. Fix Beebium.run_until_or_timeout to use server-side
cycle budget breakpoint for the timeout, with client-side predicate
polling only for the predicate itself.

48 C++ tests (189 assertions), 26 Python tests -- all passing.

+++ src/core/include/beebium/Types.hpp
@@ -65,7 +65,7 @@ enum WatchType : uint8_t {
 struct BreakpointEntry {
     uint32_t id;
     uint16_t start;
-    uint16_t end;       // exclusive [start, end)
+    uint32_t end;       // exclusive [start, end); 0x10000 for full address space
     bool stop_counterpart = false;
 
     // Condition expression (evaluated on address match).
@@ -86,7 +86,7 @@ struct BreakpointEntry {
 struct WatchpointEntry {
     uint32_t id;
     uint16_t start;
-    uint16_t end;       // exclusive [start, end)
+    uint32_t end;       // exclusive [start, end); 0x10000 for full address space
     WatchType type;     // WATCH_READ, WATCH_WRITE, WATCH_BOTH
     bool stop_counterpart = false;
